Add Ticket to Ride paper
Added link to paper analyzing Ticket To Ride strategy. Note that one of the other links in the TTR section (Applications of Graph Theory and Probability in the Board Game Ticket to Ride) is a slide deck that seems to be based largely on the paper I added. Not sure if you prefer that these two items be nested in the list or anything like that.